apps/search-base/requirements/base.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings>=2.2.1
PyYAML==6.0.2
google-cloud-aiplatform>=1.50.0
firebase-admin>=6.5.0
google-cloud-pubsub>=2.21.0
google-cloud-kms>=3.12.0
faiss-cpu>=1.7.4
numpy>=1.24.0
